requirements.txt
langchain==0.1.12
python-dotenv==1.0.1
pypdf==4.1.0
python-docx==1.1.0
streamlit==1.32.2
faiss-cpu==1.8.0
google-cloud-aiplatform==1.43.0
langchain-google-genai==0.0.11
